+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 
 ## Changed
 
-- `operator-rs` `0.12.0` -> `0.15.0` ([#137], #[153]).
+- `operator-rs` `0.12.0` -> `0.15.0` ([#137], [#153]).
+- Now using HDFS discovery config map instead of hdfs name node config map ([#153])
 
 [#133]: https://github.com/stackabletech/hbase-operator/pull/133
 [#137]: https://github.com/stackabletech/hbase-operator/pull/137